+++ packages/angular-sdk-components/src/lib/_types/PConnProps.interface.ts
@@ -1,6 +1,11 @@
 // PConnFieldProps extends PConnProps to bring in the common properties that are
 // associated with most field components (ex: Dropdown, TextInput, etc.) in the
 //  components/field directory
+export interface FieldMessageConfig {
+  content?: string;
+  visibility?: boolean;
+}
+
 export interface PConnFieldProps {
   label: string;
   required: boolean;
@@ -17,4 +22,6 @@ export interface PConnFieldProps {
   hideLabel: boolean;
   placeholder?: string;
   visibility?: boolean;
+  showFieldMessage?: boolean;
+  messageConfig?: FieldMessageConfig;
 }